--- test/image-plugin-example/src/http/get-transform-catchall/config.arc
@@ -1,5 +1,0 @@
-@aws
-runtime nodejs14.x
-memory 1152
-timeout 30
-# concurrency 1
