--- test/image-plugin-example/src/http/get-transform-catchall/index.js
@@ -1,4 +1,0 @@
-let arc = require('@architect/functions')
-let { imageHandler } = require('@ryanbethel/arc-image-plugin')
-
-exports.handler = arc.http.async(imageHandler)
